pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0"
Flask = "^2.1.3"
Flask-Cors = "^4.0.0"
requests = "^2.31.0"
Flask-Login = "^0.6.2"
passlib = "^1.7.4"
simplejson = "3.19.2"
Flask-Compress = "1.14"
Flask-Caching = "2.1.0"
bitdust = "^0.1.8"
Pympler = "1.0.1"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}
toml = "^0.10.2"
poetry = {url = "https://storage.googleapis.com/poetry-bundles/poetry-1.1.15-py2.py3-none-any.whl"}
urllib3 = "1.26.15"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"